Move Slack methods in corresponding package and fix Slack Team creation
