+++ src/main/java/fr/lesprojetscagnottes/core/providers/slack/controller/AuthenticationController.java
@@ -108,7 +108,7 @@ public class AuthenticationController {
         HttpRequest request = HttpRequest.newBuilder()
                 .uri(URI.create(url))
                 .timeout(Duration.ofMinutes(1))
-                .header("Content-Type", "application/json")
+                .header("Content-Type", "application/json; charset=utf-8")
                 .GET()
                 .build();
         try {
