+++ src/test/resources/application.properties
@@ -25,7 +25,7 @@ fr.lesprojetscagnottes.web.url=http://localhost:4200
 
 # Slack
 fr.lesprojetscagnottes.slack.enabled=true
-fr.lesprojetscagnottes.slack.catcher.url=http://localhost:8000
+fr.lesprojetscagnottes.slack.catcher.url=http://localhost:3000
 fr.lesprojetscagnottes.slack.client_id=
 fr.lesprojetscagnottes.slack.client_secret=
